requirements.txt
## Requirements for VAE

jupyterlab
numpy
scikit-image
pillow
matplotlib
opencv-python
